# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Bidly API Documentation",
        "description": "행사 기획 대행사와 용역사를 연결하는 입찰 플랫폼 API",
        "contact": {
            "name": "Bidly Support Team",
            "email": "support@bidly.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Development Server"
        }
    ],
    "paths": {
...
        "/api/notifications": {
            "get": {
                "tags": [
                    "Notifications"
                ],
                "summary": "내 알림 목록 조회",
                "description": "현재 사용자의 알림 목록을 조회합니다",
                "operationId": "911565ad9aff9ba7eded3f189509e578",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "조회할 알림 개수 (기본값: 20)",
                        "schema": {
                            "type": "integer",
                            "maximum": 100,
                            "minimum": 1
                        }
                    },
                    {
                        "name": "unread_only",
                        "in": "query",
                        "description": "읽지 않은 알림만 조회 (기본값: false)",
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "알림 목록 조회 성공",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "notifications": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Notification"
                                            }
                                        },
                                        "unread_count": {
                                            "type": "integer",
                                            "example": 5
                                        },
                                        "total_count": {
                                            "type": "integer",
                                            "example": 25
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/notifications/{notification}/read": {
            "patch": {
                "tags": [
                    "Notifications"
                ],
                "summary": "알림 읽음 처리",
                "description": "특정 알림을 읽음 처리합니다",
                "operationId": "d5a975cf47854f29f1d5bef2116b9081",
                "parameters": [
                    {
                        "name": "notification",
                        "in": "path",
                        "description": "알림 ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "알림 읽음 처리 성공",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "알림이 읽음 처리되었습니다"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "알림을 찾을 수 없습니다"
                    }
                }
            }
        },
...
        "/api/notifications/{notification}": {
            "get": {
                "tags": [
                    "Notifications"
                ],
                "summary": "알림 상세 조회",
                "description": "특정 알림의 상세 정보를 조회합니다",
                "operationId": "7a36a0e2b2148b8ade53244901d98369",
                "parameters": [
                    {
                        "name": "notification",
                        "in": "path",
                        "description": "알림 ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "알림 상세 조회 성공",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Notification"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "알림을 찾을 수 없습니다"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Notifications"
                ],
                "summary": "알림 삭제",
                "description": "특정 알림을 삭제합니다",
                "operationId": "7f078e4ee294fe830b4b842939b15a12",
                "parameters": [
                    {
                        "name": "notification",
                        "in": "path",
                        "description": "알림 ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "알림 삭제 성공",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "알림이 삭제되었습니다"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "알림을 찾을 수 없습니다"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Notification": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "user_id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "notification_type": {
                        "type": "string",
                        "example": "evaluation_step_passed"
                    },
                    "title": {
                        "type": "string",
                        "example": "1차 서류 심사 통과 안내"
                    },
                    "message": {
                        "type": "string",
                        "example": "'신년 행사' 공고의 1차 서류 심사에 통과하셨습니다."
                    },
                    "data": {
                        "type": "object"
                    },
                    "is_read": {
                        "type": "boolean",
                        "example": false
                    },
                    "read_at": {
                        "type": "string",
                        "format": "datetime",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "datetime"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "datetime"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}